requirements.txt
tkinter
sqlite3
PyQt6 
pytesseract 
pillow 
cryptography